pyproject.toml
[project]
name = "open-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.1.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.118.0",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "openai>=2.0.1",
    "pandas>=2.3.3",
    "pinecone>=7.3.0",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
